pyproject.toml
[project]
name = "kunclaw"
version = "0.1.0"
description = "KunClaw Agent Framework - Minimal, powerful, and extensible AI agent framework"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.25.0",
    "python-dotenv>=1.2.2",
]

[project.optional-dependencies]
feishu = [
    "httpx>=0.26.0",
    "lark-oapi>=1.2.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"